* adjust AA cost value - Hastened Banishment * adjusted buff duration calculation logic to be more consistent and improve splurt-style spells to be more live-like. (this is a work in progress, probably needs a few more tweaks) * fix up shadowknight proc buffs using the wrong spell id * fix up some item stacking code that was relying on the wrong item field, remove stacksize as an item property * adjusted some messages for begging skill * don't reset timer when sending cast fail message while on boat * periodic HP effect and duration counter reworked to better match eqmac client * adjusted invisibility early fade and warning message to work better with new buff duration logic * add rogue poison mastery AA tradeskill bonus * add player origin/bind locations extracted from eqmac client * reworked duel/decline client to client packet exchange

+++ zone/gm_commands/set/texture.cpp
@@ -8,7 +8,7 @@ void SetTexture(Client *c, const Seperator *sep)
 		return;
 	}
 
-	uint8 texture = Strings::ToUnsignedInt(sep->arg[2]);
+	uint16 texture = Strings::ToUnsignedInt(sep->arg[2]);
 	uint8 helmet_texture = 0xFF;
 
 	Mob *t = c;
